updated model and services
web services are not updated yet

+++ FullApp/atmsServer/src/main/java/com/atms/service/impl/DeveloperServiceImpl.java
@@ -28,7 +28,10 @@ public class DeveloperServiceImpl implements DeveloperService {
 
     @Override
     public Developer update(Developer developer) {
-        return developerRepository.saveAndFlush(developer);
+        if (developerRepository.findOne(developer.getDeveloperId()) != null) {
+            return developerRepository.saveAndFlush(developer);
+        }
+        return null;
     }
 
     @Override
@@ -42,8 +45,12 @@ public class DeveloperServiceImpl implements DeveloperService {
     }
 
     @Override
-    public void delete(Developer developer) {
-        developerRepository.delete(developer);
+    public boolean delete(Developer developer) {
+        if (developerRepository.exists(developer.getDeveloperId())) {
+            developerRepository.delete(developer);
+            return true;
+        }
+        return false;
     }
 
     @Override
